Refactor lib/auths modules structure.
This is fix #58 issue.
